CellularDataFragment shifted to StateFragment
