+++ app/src/main/java/org/oppia/app/player/state/StateFragment.kt
@@ -6,11 +6,25 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import org.oppia.app.fragment.InjectableFragment
+import org.oppia.app.player.audio.CellularDataDialogFragment
+import org.oppia.app.player.audio.CellularDataInterface
+import org.oppia.app.player.audio.LanguageInterface
 import javax.inject.Inject
+
+private const val TAG_CELLULAR_DATA_DIALOG = "CELLULAR_DATA_DIALOG"
 
 /** Fragment that represents the current state card of an exploration. */
 class StateFragment : InjectableFragment() {
-  @Inject lateinit var stateFragmentPresenter: StateFragmentPresenter
+  @Inject
+  lateinit var stateFragmentPresenter: StateFragmentPresenter
+  private lateinit var cellularDataInterface: CellularDataInterface
+  private lateinit var languageInterface: LanguageInterface
+  // Control this boolean value from controllers in domain module.
+  private var showCellularDataDialog = true
+
+  init {
+    // Add code to control the value of showCellularDataDialog using AudioController.
+  }
 
   override fun onAttach(context: Context?) {
     super.onAttach(context)
@@ -19,5 +33,37 @@ class StateFragment : InjectableFragment() {
 
   override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
     return stateFragmentPresenter.handleCreateView(inflater, container)
+  }
+
+  fun dummyButtonClicked() {
+    if (showCellularDataDialog) {
+      showCellularDataDialogFragment()
+    }
+  }
+
+  private fun showCellularDataDialogFragment() {
+    cellularDataInterface = object : CellularDataInterface {
+      override fun enableAudioWhileOnCellular(doNotShowAgain: Boolean) {
+        // Show audio-bar
+        // doNotShowAgain -> true -> save this preference
+        // doNotShowAgain -> false -> do not save this preference
+      }
+
+      override fun disableAudioWhileOnCellular(doNotShowAgain: Boolean) {
+        // Do not show audio-bar
+        // doNotShowAgain -> true -> save this preference
+        // doNotShowAgain -> false -> do not save this preference
+      }
+    }
+
+    val previousFragment = fragmentManager?.findFragmentByTag(TAG_CELLULAR_DATA_DIALOG)
+    if (previousFragment != null) {
+      fragmentManager?.beginTransaction()?.remove(previousFragment)?.commitNow()
+    }
+    val dialogFragment = CellularDataDialogFragment.newInstance(
+      cellularDataInterface
+    )
+    // Might need to replace show() with showNow().
+    dialogFragment.showNow(fragmentManager, TAG_CELLULAR_DATA_DIALOG)
   }
 }
